app/pyproject.toml
[tool.poetry]
name = "Ai-powered-jenkins-auditor"
version = "0.1.0"
description = ""
authors = ["taSherb22 <mohamedtaher.boudrigua@supcom.tn>"]
readme = "README.md"
packages = [{include = "ai_powered_jenkins_auditor"}]


[tool.poetry.dependencies]
python = "3.11.9"
tree-sitter = "^0.24.0"
tree-sitter-groovy = "^0.1.2"
flask = "^3.1.1"
pyngrok = "^7.2.12"
python-dotenv = "^1.1.1"
presidio-analyzer = "^2.2.359"
presidio-anonymizer = "^2.2.359"
gunicorn = "^23.0.0"
langchain-core = "^0.3.72"
pydantic = "^2.11.7"
langchain-community = "^0.3.27"
langgraph = "^0.6.0"
langchain-groq = "^0.3.6"
langchain-tavily = "^0.2.11"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
python-multipart = "^0.0.20"
watchdog = "^6.0.0"





[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-xdist = "^3.8.0"

[build-system]

build-backend = "poetry.core.masonry.api"
